backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
websockets>=13.0,<16
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.12
openai==1.50.0
groq==0.11.0
aiofiles==23.2.1
supabase==2.10.0